requirements.txt
appnope==0.1.0
astroid==1.6.1
attrs==17.4.0
autopep8==1.3.4
bleach==2.1.2
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isort==4.3.4
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
Pygments==2.2.0
pylint==1.8.2
pyparsing==2.2.0
pyquaternion==0.9.2
pytest==3.4.0
python-dateutil==2.6.1
pytz==2018.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
torch==0.3.1.post2
torchfile==0.1.0
torchvision==0.2.0
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
visdom==0.1.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
wrapt==1.10.11
